chore: move utils to file
