+++ src/index.ts
@@ -1,6 +1,11 @@
 import fs from 'fs';
 import path from 'path';
 import yargs from 'yargs';
+import {
+  getTargetFolder,
+  maybeCreateTargetFolder,
+  retrieveFiles,
+} from './utils';
 import { renameToNewFile } from './utils/rename';
 
 interface ExpectedArguments {
@@ -42,54 +47,6 @@ const parsedArgs: ExpectedArguments = yargs(process.argv.slice(2))
     'Show help. See https://github.com/siuangie91/batch-renamer#batch-renamer'
   )
   .parse(process.argv.slice(2));
-
-/**
- * Creates the absolute path to the target folder.
- * If no target was provided by the user,
- * appends `_renamed` to the original folder name
- * and uses that as the target folder name.
- * @param props
- * @returns
- */
-const getTargetFolder = ({
-  target,
-  originFolderName,
-  originParent,
-}: {
-  target: string | null;
-  originFolderName: string;
-  originParent: string;
-}): string => {
-  // if not target path not provided,
-  // use original name with `_renamed` appended
-  const backupTargetFolderName = `${originFolderName}_renamed`;
-
-  // use backup target folder if target not provided
-  const targetFolder = target || `${originParent}/${backupTargetFolderName}`;
-
-  return targetFolder;
-};
-
-/**
- * Creates the target folder if it doesn't already exist
- * @param targetFolder absolute path to the target folder
- */
-const maybeCreateTargetFolder = (targetFolder: string): void => {
-  if (!fs.existsSync(targetFolder)) {
-    fs.mkdirSync(targetFolder);
-    console.log('🛠 Created target folder', targetFolder);
-  }
-};
-
-const retrieveFiles = (origin: string): string[] => {
-  const files = fs.readdirSync(origin);
-
-  if (!files.length) {
-    throw new Error(`❌ Failed to read origin folder at path: ${origin}`);
-  }
-
-  return files;
-};
 
 const batchRename = (args: ExpectedArguments): void => {
   const { origin, prefix, target, startingIndex } = args;
